Helpers now working, general methods also working
